automatically backed up by learn

+++ lib/create.sql
@@ -2,7 +2,7 @@ CREATE TABLE projects (
   id INTEGER PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal REAL,
+  funding_goal INTEGER,
   start_date TEXT,
   end_date TEXT
 );
